add actions for location validation and rename reducer files
