+++ client/src/actions/types.js
@@ -21,3 +21,7 @@ export const SELECT_PROJECT_POST = 'select_project_post';
 // MAP DATA
 export const RECEIVE_MAP_DATA = 'receive_map_data';
 export const REQUEST_MAP_DATA = 'request_map_data';
+
+// Validate user post;
+export const REQUEST_LOCATION = 'request_location';
+export const RECEIVE_LOCATION = 'receive_location';
