+++ client/src/reducers/mapDataReducer.js
